--- src/main/java/roomescape/exception/DuplicatedReservationTimeException.java
@@ -1,9 +1,0 @@
-package roomescape.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class DuplicatedReservationTimeException extends CustomException {
-    public DuplicatedReservationTimeException() {
-        super("중복된 예약시간입니다.", HttpStatus.CONFLICT);
-    }
-}
