+++ src/main/java/roomescape/exception/reservation/InvalidDateTimeReservationException.java
@@ -1,6 +1,7 @@
-package roomescape.exception;
+package roomescape.exception.reservation;
 
 import org.springframework.http.HttpStatus;
+import roomescape.exception.CustomException;
 
 public class InvalidDateTimeReservationException extends CustomException {
     public InvalidDateTimeReservationException() {
