+++ src/main/java/roomescape/exception/time/ReservationReferencedTimeException.java
@@ -1,6 +1,7 @@
-package roomescape.exception;
+package roomescape.exception.time;
 
 import org.springframework.http.HttpStatus;
+import roomescape.exception.CustomException;
 
 public class ReservationReferencedTimeException extends CustomException {
     public ReservationReferencedTimeException() {
